FFT Widget Overhaul  Committer: JP Ryan <jryan@AVIRIS-CAL.jpl.nasa.gov> 	modified:   Makefile Modified the Makefile for cuda take to include information from the command line on the compiler author and date/time of compilation to help with version control. This will display as the first lines of liveview when the program is first started.
	modified:   include/constants.h
Added a constant for the TAP_WIDTH. Taps are assumed to be 160 pixels wide.

	modified:   src/mean_filter.cpp
Updated the mean filter to interpret different types of FFT inputs.
We now have FFTs of not just the frame mean, but also of a vertical
crosshair, and of an entire tap.

	modified:   src/take_object.cpp
take_object has been modified to include information on the version author and modification time.
I also added functions which mediate information between liveview and the mean filter.