FFT Widget Overhaul  Committer: JP Ryan <jryan@AVIRIS-CAL.jpl.nasa.gov> 	modified:   controlsbox.cpp Updated the controls for the FFT widget to control for situations which would mess up the data or cause seg faults. The LinesToAverage slider needs to be active only when the v. crosshair radio button in the FFT widget is checked.
	modified:   fft_widget.cpp
Added controls to specify which type of FFT you would like to take.
This includes a spinBox for the tap to profile.
The xAxis of the bar graph changes based on the Nyquist Frequency of
the input type:
Frame Mean: Frame Frequency / 2 = ~60Hz
Vertical Crosshair: Row Frequency / 2 = ~28kHz
Tap Profile: Pixel Frequency / 2 = ~5MHz

In the mode for the vertical crosshair, you may choose how many
columns  you would like to average (reduce noise)

	modified:   liveview2.pro
Added to the makefile to define the version author and modification time