Added ZUpT disable to the GPS Manager.
ZUpT will be disabled upon successful GPS TCP connection. It is sent
immediately as well as three seconds later. If the GPS is lost and then
reconnected, it is sent again and again three seconds later.

Currently, we are not sending this out when we detect a zupt issue, but
we might add that later.